chore: split up pox-2-rosetta tests
