+++ .github/workflows/ci.yml
@@ -411,7 +411,9 @@ jobs:
           pox-2-delegate-aggregation,
           pox-2-delegate-stacking,
           pox-2-stack-extend-increase,
-          pox-2-rosetta
+          pox-2-rosetta-btc-addr-types,
+          pox-2-rosetta-cycle-phases,
+          pox-2-rosetta-segwit,
         ]
     runs-on: ubuntu-latest
     steps:
